# ndvi/docs/openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","description":"NDVI map constructor","version":"0.1.0"},"paths":{"/api/upload_geo":{"post":{"summary":"Add Geojson","operationId":"add_geojson_api_upload_geo_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_add_geojson_api_upload_geo_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/show_fields":{"get":{"summary":"Get Fields List","operationId":"get_fields_list_api_show_fields_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/delete_field/{field_id}":{"delete":{"summary":"Delete Data","operationId":"delete_data_api_delete_field__field_id__delete","parameters":[{"required":true,"schema":{"title":"Field Id","type":"integer"},"name":"field_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/download/{field_id}":{"get":{"summary":"Get Ndvi Image","operationId":"get_ndvi_image_api_download__field_id__get","parameters":[{"required":true,"schema":{"title":"Field Id","type":"integer"},"name":"field_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/start_processing/":{"post":{"summary":"Start Processing Datasets","operationId":"start_processing_datasets_api_start_processing__post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_add_geojson_api_upload_geo_post":{"title":"Body_add_geojson_api_upload_geo_post","required":["file","name"],"type":"object","properties":{"file":{"title":"File","type":"string","format":"binary"},"name":{"title":"Name","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}